Add a real-time STFT spectrogram
This is done to observe how the spectrum changes when a string is struck at different velocities.
The STFT is implemented in the SpectrogramComponent class. It is initialized in the PluginProcessor, which the pushes the currently synthesized samples to a FIFO. The SpectrogramComponent has a timer, whose callback computes and stores an STFT line into a juce::Image. This image is then retrieved and painted by the PluginEditor in OpenPianoAudioProcessorEditor::paint.
As a side note, rectangles containing bounds are now used instead of the juce::Grid. This is done to simplify the co-existence of the piano keyboard with a painted component (the STFT).