Implement a harmonic ratio visualization
The harmonic ratio measures the amount of energy in the tonal part of the signal compared to the amount of energy in the total signal. While the STFT lets us see how the spectrum becomes richer as we hit harder the string, the harmonic ratio shows how much inharmonicity there is depending on the string.